HandwerkerImperium: Feature-Integration + Bugfixes in bestehenden Models und Services
Models: DailyChallenge int→long, Order.CalculateEstimatedReward() zentralisiert,
Tool 4 neue Typen (Hammer/Wasserwaage/Lupe/Zirkel), GameState um neue Feature-Felder erweitert,
Prestige 7 Stufen (Bronze bis Legende).

Services: GameLoopService neue Feature-Ticks, AchievementService Gebäude-Cases,
OfflineProgress Saison-Multiplikator-Fix, StoryService skalierte Belohnungen,
PlayGamesService Leaderboard-Integration, DI-Registrierung aller neuen Services.

+++ src/Apps/HandwerkerImperium/HandwerkerImperium.Shared/Models/DailyChallenge.cs
@@ -28,10 +28,10 @@ public class DailyChallenge
     public DailyChallengeType Type { get; set; }
 
     [JsonPropertyName("targetValue")]
-    public int TargetValue { get; set; }
+    public long TargetValue { get; set; }
 
     [JsonPropertyName("currentValue")]
-    public int CurrentValue { get; set; }
+    public long CurrentValue { get; set; }
 
     [JsonPropertyName("moneyReward")]
     public decimal MoneyReward { get; set; }
